+++ ccramic/callbacks/pixel_level_callbacks.py
@@ -36,7 +36,10 @@ from ccramic.utils.pixel_level_utils import (
 # from ccramic.utils.cell_level_utils import generate_greyscale_grid_array
 # from ccramic.utils.session import remove_ccramic_caches
 from ccramic.components.canvas import CanvasImage, CanvasLayout
-from ccramic.io.display import generate_area_statistics_dataframe, output_current_canvas_as_tiff
+from ccramic.io.display import (
+    generate_area_statistics_dataframe,
+    output_current_canvas_as_tiff,
+    output_current_canvas_as_html)
 from ccramic.io.gallery_outputs import generate_channel_tile_gallery_children
 from ccramic.parsers.cell_level_parsers import match_mask_name_with_roi
 from ccramic.utils.graph_utils import strip_invalid_shapes_from_graph_layout
@@ -48,7 +51,8 @@ from ccramic.callbacks.pixel_level_wrappers import parse_global_filter_values_fr
 from ccramic.io.session import (
     write_blend_config_to_json,
     write_session_data_to_h5py,
-    subset_mask_for_data_export)
+    subset_mask_for_data_export,
+    create_download_dir)
 # from ccramic.parsers.cell_level_parsers import validate_coordinate_set_for_image
 from pathlib import Path
 from plotly.graph_objs.layout import YAxis, XAxis
@@ -831,15 +835,14 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
 
     @dash_app.callback(
         Input('data-collection', 'value'),
-        Output('canvas-layers', 'data', allow_duplicate=True),
-        Output("download-collapse", "is_open"))
+        Output('canvas-layers', 'data', allow_duplicate=True))
     def reset_canvas_layers_on_new_dataset(data_selection):
         """
         Reset the canvas layers dictionary containing the cached images for the current canvas in order to
         retain memory. Should be cleared on a new ROI selection
         """
         if data_selection is not None:
-            return None, False
+            return None
         else:
             raise PreventUpdate
 
@@ -1477,93 +1480,72 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         else:
             raise PreventUpdate
 
-    @dash_app.callback(Output('download-link', 'href'),
-                       Output('download-canvas-interactive-html', 'href'),
-                       Output('download-blend-config', 'href'),
-                       State('uploaded_dict', 'data'),
-                       State('imc-metadata-editable', 'data'),
-                       State('blending_colours', 'data'),
-                       Input("open-download-collapse", "n_clicks"),
-                       Input("download-collapse", "is_open"),
-                       State('data-collection', 'value'),
+    @dash_app.callback(Output('download-canvas-image-html', 'data'),
+                       Input('btn-download-canvas-html', 'n_clicks'),
                        State('annotation_canvas', 'figure'),
+                       State('uploaded_dict', 'data'),
+                       State('blending_colours', 'data'),
+                       State('annotation_canvas', 'style'))
+    # @cache.memoize())
+    def download_interactive_html_canvas(download_html, cur_graph, uploaded, blend_dict, canvas_style):
+        if None not in (cur_graph, uploaded, blend_dict) and download_html > 0:
+            try:
+                download_dir = os.path.join(tmpdirname, authentic_id, 'downloads')
+                create_download_dir(download_dir)
+                html_path = dcc.send_file(output_current_canvas_as_html(cur_graph, canvas_style, download_dir))
+            except (ValueError, KeyError):
+                html_path = dash.no_update
+            return html_path
+        else:
+            raise PreventUpdate
+
+    @dash_app.callback(Output('download-session-config-json', 'data'),
+                       Input('btn-download-session-config-json', 'n_clicks'),
+                       State('blending_colours', 'data'),
                        State('image_layers', 'value'),
-                       State('annotation_canvas', 'style'),
-                       State('annotation_canvas', 'relayoutData'),
-                       State('graph-subset-download', 'value'),
                        State('bool-apply-global-filter', 'value'),
                        State('global-filter-type', 'value'),
                        State("global-kernel-val-filter", 'value'),
                        State("global-sigma-val-filter", 'value'))
     # @cache.memoize())
-    def update_download_href_h5(uploaded, metadata_sheet, blend_dict, nclicks, download_open, data_selection,
-                                current_canvas, blend_layers, canvas_style, canvas_layout, graph_subset,
+    def download_session_config_json(download_json, blend_dict, blend_layers,
                                 global_apply_filter, global_filter_type, global_filter_val, global_filter_sigma):
+        if blend_dict is not None and download_json > 0:
+            download_dir = os.path.join(tmpdirname, authentic_id, 'downloads')
+            create_download_dir(download_dir)
+            return dcc.send_file(write_blend_config_to_json(download_dir, blend_dict, blend_layers, global_apply_filter,
+                                                global_filter_type, global_filter_val, global_filter_sigma))
+        else:
+            raise PreventUpdate
+
+    @dash_app.callback(Output('download-roi-h5py', 'data'),
+                       # Output('download-canvas-interactive-html', 'href'),
+                       Input('btn-download-roi-h5py', 'n_clicks'),
+                       State('uploaded_dict', 'data'),
+                       State('imc-metadata-editable', 'data'),
+                       State('blending_colours', 'data'),
+                       State('data-collection', 'value'),
+                       State('annotation_canvas', 'relayoutData'),
+                       State('graph-subset-download', 'value'))
+    # @cache.memoize())
+    def update_download_href_h5(download_h5py, uploaded, metadata_sheet, blend_dict, data_selection,
+                                canvas_layout, graph_subset):
         """
         Create the download links for the current canvas and the session data.
         Only update if the download dialog is open to avoid continuous updating on canvas change
         """
-        if None not in (uploaded, blend_dict) and nclicks > 0 and download_open:
-
+        if None not in (uploaded, blend_dict) and download_h5py > 0:
             first_image = get_first_image_from_roi_dictionary(uploaded[data_selection])
             download_dir = os.path.join(tmpdirname, authentic_id, 'downloads')
-            if not os.path.exists(download_dir):
-                os.makedirs(download_dir)
-            # TODO: create testable function for the download mask
+            create_download_dir(download_dir)
             try:
                 mask = None
                 if 'shapes' in canvas_layout and ' use graph subset on download' in graph_subset:
                     mask = subset_mask_for_data_export(canvas_layout, first_image.shape)
-
-                relative_filename = write_session_data_to_h5py(download_dir, metadata_sheet, uploaded,
-                                                               data_selection, blend_dict, mask)
-                if not ' use graph subset on download' in graph_subset:
-                    try:
-                        fig = go.Figure(current_canvas)
-                        # fig.update_layout(xaxis_showgrid=False, yaxis_showgrid=False,
-                        #                   xaxis=XAxis(showticklabels=False),
-                        #                   yaxis=YAxis(showticklabels=False),
-                        #                   margin=dict(l=0, r=0, b=0, t=0, pad=0))
-                        fig.update_layout(dragmode="zoom")
-                        fig.write_html(str(os.path.join(download_dir, "canvas.html")),
-                            default_width = canvas_style['width'], default_height = canvas_style['height'])
-                        fig_return = str(os.path.join(download_dir, "canvas.html"))
-                    except ValueError:
-                        fig_return = dash.no_update
-                else:
-                    fig_return = dash.no_update
-                # param_json = str(os.path.join(download_dir, 'param.json'))
-                # with open(param_json, "w") as outfile:
-                #     dict_write = {"channels": blend_dict, "config":
-                #         {"blend": blend_layers, "filter":
-                #         {"global_apply_filter": global_apply_filter, "global_filter_type": global_filter_type,
-                #          "global_filter_val": global_filter_val, "global_filter_sigma": global_filter_sigma}}}
-                #     json.dump(dict_write, outfile)
-                param_json = write_blend_config_to_json(download_dir, blend_dict, blend_layers, global_apply_filter,
-                               global_filter_type, global_filter_val, global_filter_sigma)
-
-                return str(relative_filename), fig_return, param_json
+                return dcc.send_file(write_session_data_to_h5py(download_dir, metadata_sheet, uploaded,
+                                                               data_selection, blend_dict, mask))
             # if the dictionary hasn't updated to include all the experiments, then don't update download just yet
             except KeyError:
-                raise PreventUpdate
-        else:
-            raise PreventUpdate
-
-    @dash_app.callback(
-        Input('current_canvas_image', 'data'),
-        Input('annotation_canvas', 'figure'),
-        State("download-collapse", "is_open"),
-        Input('annotation_canvas', 'relayoutData'),
-        Output("download-collapse", "is_open", allow_duplicate=True),
-        prevent_initial_call=True)
-    def reset_canvas_layers_on_new_dataset(current_image, current_canvas, currently_open, canvas_layout):
-        """
-        Close the collapsible download when an update is made to the canvas to prevent extraneous downloading
-        """
-        if None not in (current_canvas, canvas_layout):
-            if currently_open:
-                return False
-            else:
                 raise PreventUpdate
         else:
             raise PreventUpdate
@@ -1758,16 +1740,6 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         else:
             return pd.DataFrame({}, columns=["Channel"]).to_dict(orient="records"), \
                         {"sortable": False, "filter": False}
-
-    @dash_app.callback(
-        Output("download-collapse", "is_open", allow_duplicate=True),
-        [Input("open-download-collapse", "n_clicks")],
-        [State("download-collapse", "is_open")])
-    # @cache.memoize())
-    def toggle_download_collapse(n, is_open):
-        if n:
-            return not is_open
-        return is_open
 
     @dash_app.callback(
         Output("area-stats-collapse", "is_open", allow_duplicate=True),