+++ tests/test_display_outputs.py
@@ -2,10 +2,13 @@ import os.path
 
 from ccramic.io.display import (
     generate_area_statistics_dataframe,
-    output_current_canvas_as_tiff)
+    output_current_canvas_as_tiff,
+    output_current_canvas_as_html)
 import numpy as np
 import pandas as pd
 import tempfile
+import plotly.graph_objs as go
+import plotly.express as px
 
 def test_generate_channel_statistics_dataframe():
     upload_dict = {"experiment0+++slide0+++acq0": {"DNA": np.full((1000, 1000), 100),
@@ -237,6 +240,20 @@ def test_output_canvas_tiff_to_file():
         file_path = os.path.join(tmpdirname, "canvas.tiff")
         assert not os.path.exists(file_path)
         canvas_link = output_current_canvas_as_tiff(canvas_image, tmpdirname)
+        assert os.path.exists(file_path)
+        assert str(file_path) == str(canvas_link)
+        if os.access(canvas_link, os.W_OK):
+            os.remove(canvas_link)
+        assert not os.path.exists(canvas_link)
+
+def test_output_canvas_html_to_file():
+    canvas_image = np.full((100, 100, 3), 7)
+    fig = go.Figure(px.imshow(canvas_image)).to_dict()
+    style = {"height": 100, "width": 100}
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        file_path = os.path.join(tmpdirname, "canvas.html")
+        assert not os.path.exists(file_path)
+        canvas_link = output_current_canvas_as_html(fig, style, tmpdirname)
         assert os.path.exists(file_path)
         assert str(file_path) == str(canvas_link)
         if os.access(canvas_link, os.W_OK):